Organize package structure and refactor classes in the order_item package
