+++ src/main/java/com/ecommerce/api/order/controller/OrderController.java
@@ -3,7 +3,7 @@ package com.ecommerce.api.order.controller;
 import com.ecommerce.api.order.request.OrderRequest;
 import com.ecommerce.api.order.response.OrderResponse;
 import com.ecommerce.api.order.service.OrderService;
-import com.ecommerce.api.order_item.RequestOrderItemDto;
+import com.ecommerce.api.order_item.request.OrderItemRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -26,8 +26,8 @@ public class OrderController {
     }
 
     @PostMapping("/addToOrder")
-    public ResponseEntity<OrderResponse> addToCart(@RequestBody RequestOrderItemDto requestOrderItemDto){
-        OrderResponse responseDto = orderService.addToOrder(requestOrderItemDto);
+    public ResponseEntity<OrderResponse> addToCart(@RequestBody OrderItemRequest orderItemRequest){
+        OrderResponse responseDto = orderService.addToOrder(orderItemRequest);
         return ResponseEntity.ok(responseDto);
     }
