+++ src/main/java/com/ecommerce/api/order/model/Order.java
@@ -1,6 +1,6 @@
 package com.ecommerce.api.order.model;
 
-import com.ecommerce.api.order_item.OrderItem;
+import com.ecommerce.api.order_item.model.OrderItem;
 import com.ecommerce.api.utility.EnumConstants;
 import com.ecommerce.api.order_timeline.OrderTimeline;
 import com.ecommerce.api.user.User;
