+++ src/main/java/com/ecommerce/api/order_item/service/OrderItemRepository.java
@@ -1,5 +1,6 @@
-package com.ecommerce.api.order_item;
+package com.ecommerce.api.order_item.service;
 
+import com.ecommerce.api.order_item.model.OrderItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
